requirements.txt
alembic==1.13.3
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
fastapi==0.114.0
fastapi-security==0.5.0
Flask==3.0.3
Flask-Cors==5.0.0
flask-restplus==0.13.0
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2==2.9.9
pydantic==1.10.18
pydantic_core==2.23.2
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.4
swagger_ui==0.1.2
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
websockets==13.0.1
Werkzeug==3.0.4
